wip: simplify component
