+++ package.json
@@ -16,7 +16,7 @@
     },
     "dependencies": {
         "@floating-ui/react": "^0.27.16",
-        "@lottiefiles/dotlottie-react": "^0.15.1",
+        "@lottiefiles/dotlottie-react": "^0.15.2",
         "@number-flow/react": "^0.5.9",
         "@tanstack/react-query": "^5.85.5",
         "@tari-project/tari-tower": "^0.1.4",
@@ -52,11 +52,11 @@
         "@taplo/cli": "^0.7.0",
         "@tauri-apps/cli": "^2.8.3",
         "@types/node": "^24.3.0",
-        "@types/react": "^19.1.11",
+        "@types/react": "^19.1.12",
         "@types/react-dom": "^19.1.8",
         "@types/uuid": "^10.0.0",
         "@typescript-eslint/parser": "^8.3.0",
-        "@vitejs/plugin-react": "^5.0.1",
+        "@vitejs/plugin-react": "^5.0.2",
         "babel-plugin-react-compiler": "^19.1.0-rc.2",
         "babel-plugin-styled-components": "^2.1.4",
         "dotenv": "^17.2.1",
